+++ bundle-es2015/entries/entry-operations.js
@@ -1,15 +1,16 @@
 import { LinkResolver } from './link-resolver';
-import { defaultMapperForLanguage, defaultMapperForPublishedVersionStatus, isBrowser, isIE, UrlBuilder } from 'contensis-core-api';
+import { defaultMapperForLanguage, defaultMapperForPublishedVersionStatus, isBrowser, isIE, Query, UrlBuilder, ZenqlQuery } from 'contensis-core-api';
+const defaultListUrl = `/api/delivery/projects/:projectId/entries`;
+let listUrl = (options, params) => {
+    return !!options.contentTypeId
+        ? `/api/delivery/projects/:projectId/contentTypes/:contentTypeId/entries`
+        : defaultListUrl;
+};
 let getMappers = {
     language: defaultMapperForLanguage,
     versionStatus: defaultMapperForPublishedVersionStatus,
     fields: (value) => (value && value.length > 0) ? value : null,
     linkDepth: (value) => (value && (value > 0)) ? value : null,
-};
-let listUrl = (options, params) => {
-    return !!options.contentTypeId
-        ? `/api/delivery/projects/:projectId/contentTypes/:contentTypeId/entries`
-        : `/api/delivery/projects/:projectId/entries`;
 };
 let listMappers = {
     ...getMappers,
@@ -45,20 +46,77 @@ export class EntryOperations {
         if (!query) {
             return new Promise((resolve) => { resolve(null); });
         }
+        let deliveryQuery = query instanceof Query ? query : null;
+        // use duck-typing for backwards compatibility pre v1.2.0
+        if (deliveryQuery !== null || !!query.where || !!query.orderBy) {
+            return this.searchUsingQuery(deliveryQuery || query, linkDepth);
+        }
+        let zenqlQuery = query instanceof ZenqlQuery ? query : null;
+        if (zenqlQuery === null) {
+            if (typeof query === 'string') {
+                zenqlQuery = new ZenqlQuery(query);
+            }
+            else {
+                throw new Error('A valid query needs to be specified.');
+            }
+        }
         let params = this.paramsProvider.getParams();
-        let pageSize = query.pageSize || params.pageSize;
-        let pageIndex = query.pageIndex || 0;
-        let orderBy = (query.orderBy && (query.orderBy._items || query.orderBy));
+        let pageSize = params.pageSize || 25;
+        let pageIndex = params.pageIndex || 0;
+        let fields = [];
+        pageSize = zenqlQuery.pageSize || pageSize;
+        pageIndex = zenqlQuery.pageIndex || pageIndex;
+        fields = zenqlQuery.fields || fields;
         let { accessToken, projectId, language, responseHandler, rootUrl, versionStatus, ...requestParams } = params;
         let payload = {
             ...requestParams,
             linkDepth,
             pageSize,
             pageIndex,
-            fields: query.fields && query.fields.length > 0 ? query.fields : null,
-            where: JSON.stringify(query.where),
+            zenql: zenqlQuery.zenql,
         };
-        if (orderBy && orderBy.length > 0) {
+        if (fields && fields.length > 0) {
+            payload['fields'] = fields;
+        }
+        let url = UrlBuilder.create(defaultListUrl, { ...payload })
+            .setParams({ ...payload, projectId })
+            .addMappers(searchMappers)
+            .toUrl();
+        return this.httpClient.request(url, {
+            method: 'GET',
+            headers: { 'Content-Type': 'application/json; charset=utf-8' }
+        });
+    }
+    resolve(entryOrList, fields = null) {
+        let params = this.paramsProvider.getParams();
+        let resolver = new LinkResolver(entryOrList, fields, params.versionStatus, (query) => this.search(query));
+        return resolver.resolve();
+    }
+    searchUsingQuery(query, linkDepth = 0) {
+        if (!query) {
+            return new Promise((resolve) => { resolve(null); });
+        }
+        let deliveryQuery = query;
+        let params = this.paramsProvider.getParams();
+        let pageSize = params.pageSize || 25;
+        let pageIndex = params.pageIndex || 0;
+        let fields = [];
+        pageSize = deliveryQuery.pageSize || pageSize;
+        pageIndex = deliveryQuery.pageIndex || pageIndex;
+        fields = deliveryQuery.fields || fields;
+        let orderBy = (deliveryQuery.orderBy && (deliveryQuery.orderBy._items || deliveryQuery.orderBy));
+        let { accessToken, projectId, language, responseHandler, rootUrl, versionStatus, ...requestParams } = params;
+        let payload = {
+            ...requestParams,
+            linkDepth,
+            pageSize,
+            pageIndex,
+            where: JSON.stringify(deliveryQuery.where)
+        };
+        if (fields && fields.length > 0) {
+            payload['fields'] = fields;
+        }
+        if (deliveryQuery.orderBy && (!Array.isArray(deliveryQuery.orderBy) || deliveryQuery.orderBy.length > 0)) {
             payload['orderBy'] = JSON.stringify(orderBy);
         }
         let url = UrlBuilder.create('/api/delivery/projects/:projectId/entries/search', { ...payload })
@@ -72,11 +130,6 @@ export class EntryOperations {
             method: 'GET',
             headers: { 'Content-Type': 'application/json; charset=utf-8' }
         });
-    }
-    resolve(entryOrList, fields = null) {
-        let params = this.paramsProvider.getParams();
-        let resolver = new LinkResolver(entryOrList, fields, params.versionStatus, (query) => this.search(query));
-        return resolver.resolve();
     }
     searchUsingPost(query, linkDepth = 0) {
         if (!query) {