+++ lib/index.d.ts
@@ -1,2 +1,2 @@
-export { Query, Op, OrderBy } from 'contensis-core-api';
+export { Op, OrderBy, Query, ZenqlQuery } from 'contensis-core-api';
 export { Client } from './client/client';