+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2022-06-08
+### Changes
+-  entries: changed *entries.search* to use either a *string*, a *Query* or a *ZenqlQuery* type
+
 ## [1.1.9] - 2022-04-14
 ### Changes
-- Entry.sys property is now a Partial
-- updated optional and required fields for EntrySys class
+- *Entry.sys* property is now a *Partial*
+- updated optional and required fields for *EntrySys* class
 
 ## [1.1.8] - 2021-10-15
 ### Added